Don't set content type for internal users autocomplete calls
Besides AJAX calls, /Helpers/Autocomplete/Users is also used internally
to get user info for selectize default options(in /Elements/EmailInput).
Thus always setting json http response content type is inappropriate.
$abort could be used to distinguish these 2 usages as we only set $abort
to 0 for internal usages.

This is not an issue in most cases as we flush buffer in
/Elements/PageLayout, where text/html content type is already sent to
browsers in advance.  This is to get around issues for customized pages
without /Elements/PageLayout or an explicit flush_buffer call.